requirements.txt
anthropic>=0.40.0
beautifulsoup4>=4.12.0
fastapi>=0.115.0
httpx>=0.27.0
playwright>=1.48.0
pydantic>=2.9.0
pymongo>=4.10.0
python-dotenv>=1.0.1
python-telegram-bot>=21.0
uvicorn>=0.32.0
